Combine some filesystemutils for windows and mac together.
Formatting.
